requirements.txt
Flask==2.3.2
Flask_RESTful==0.3.10
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.16
